feat: restyle navbar component
